pyproject.toml
[project]
name = "marlene"
version = "0.1.0"
description = "Custom smart home assistant with Deepgram voice agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyaudio>=0.2.14",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pvporcupine>=3.0.0",
    "deepgram-sdk>=3.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]
